update to "proper" julia package structure
- from code-pair session
