+++ Project.toml
@@ -1,8 +1,20 @@
+name = "TETB"
+uuid = "0d4be8d3-4503-476c-b33f-e073dd414f32"
+authors = ["Antonio Morales Perez <antonio.morales@dipc.org> and contributors"]
+version = "0.0.1"
+
 [deps]
 Crystalline = "ae5e2be0-a263-11e9-351e-f94dad1eb351"
-Gurobi = "2e9cd046-0924-5485-92f1-d5272153d98b"
-JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 MPBUtils = "a5feb4ea-5495-4677-96f8-dddb76c9eb4c"
 PhotonicBandConnectivity = "d7bc101a-f969-4292-96d7-1ebdabfa903f"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 SymmetryBases = "cb68ded9-7e8c-45d3-bc66-19b5867c0467"
+
+[compat]
+julia = "1.8"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
