+++ README.md
@@ -1,0 +1,3 @@
+# TETB.jl
+
+[![Build Status](https://github.com/AntonioMoralesPerez/TETB.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/AntonioMoralesPerez/TETB.jl/actions/workflows/CI.yml?query=branch%3Amain)
