+++ examples/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+TETB = "0d4be8d3-4503-476c-b33f-e073dd414f32"
