+++ test/runtests.jl
@@ -1,0 +1,6 @@
+using TETB
+using Test
+
+@testset "TETB.jl" begin
+    # Write your tests here.
+end
